rename: UserService to UserInputSerivce and move files
GameHandler를 Application class가 있는 위치로 변경.
UserService 를 UserInputService로 이름을 바꾼 뒤 service package로 이동.
